requirements.txt
mysql-connector-python, import mysql.connector
requests
python-dotenv
pyttsx3
os